Extract RSolr::Field into a separate file
